Extract shared tag/room filter logic; wire calType/mode + zoom-room filtering into WeeklyView
